==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(url=URL)
    
    if response.status_code == 200:
        path = f"../data/raw/rides_{year}-{month:02d}.parquet"
        with open(path, "wb") as f:
            f.write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available.")

download_one_file_of_raw_data(year=2022, month=1)

import pandas as pd

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")

# Only need when and where the pickup happened
rides.head()

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(
    columns={
        'tpep_pickup_datetime': 'pickup_datetime', 
        'PULocationID': 'pickup_location_id',
    }, inplace=True
)

rides.head()

### Validate dataframe

rides['pickup_datetime'].describe()

- Data should be just in January 2022, but we have one entry from 2008 and another from May 2022. 
- Remove any rows outside of Jan 22.

mask = (rides['pickup_datetime'] >= '2022-01-01') & (rides['pickup_datetime'] < '2022-02-01')

rides = rides[mask]

rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
