==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet("../data/transformed/validated_rides_2022_01.parquet")
rides.head()

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('h')
rides

agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)
agg_rides

Problem: groupby throws away rows where there is a count of zero. In our case, if there is a (location, hour) pair that had no pickups, these will have been thrown away. 

- This might not have been the case in NYC since there are a lot of people and a lot of taxis, but for other applications, e.g. eCommerce, we could easily have very sparse time series for products that have low demand. 
- We'll fix this so that we have complete time series for every location. 


from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    
    location_ids = agg_rides['pickup_location_id'].unique()
    # Makde a full range of datetime indexes.
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='h'
    )
    # For storing the modified time series'
    output = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        
        # Temporary df holding rides with current location id
        tmp = agg_rides.loc[agg_rides['pickup_location_id'] == location_id, ['pickup_hour', 'rides']]
        
        # Quick way of adding missing dates and filling with 0 in a Series; from https://stackoverflow.com/a/19324591
        tmp.set_index('pickup_hour', inplace=True)
        tmp.index = pd.DatetimeIndex(tmp.index)
        tmp = tmp.reindex(full_range, fill_value=0)
        
        # Add back in the location id
        tmp['pickup_location_id'] = location_id
        
        output = pd.concat([output, tmp])
    
    # Move purchase date from index to column
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

from typing import Optional, List
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
):
    rides_to_plot = rides[rides['pickup_location_id'].isin(locations)] if locations else rides
    
    fig = px.line(
        rides_to_plot,
        x='pickup_hour',
        y='rides',
        color='pickup_location_id',
        template='none',
    )
    
    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
