==> notebooks/03_transform_ts_data_into_features_and_target.ipynb <==
# TODO: I think this could all be done more efficiently using feature-engine, e.g. WindowFeatures, etc. 
#   Check this out once I've done it his way

# TODO: Check out paralell processing using: joblib, Ray, Dask. For loops are slow in python so this is worth doing for
#   bigger jobs

import pandas as pd

ts_data = pd.read_parquet("../data/transformed/ts_data_2022_01.parquet")
ts_data.head()

# Just slice for a single location at first.
# Not sure we need the reset_index here (see next cell)
ts_data_one_location = ts_data.loc[ts_data['pickup_location_id'] == 43, :].reset_index(drop=True)
ts_data_one_location.head()

ts_data.loc[ts_data['pickup_location_id'] == 43, :]

- We'll obtain the features for training by choosing a window size, e.g. 24 hrs, and using these values to predict the demand for the next hour. 
- To produce the entire training set, we write a function that slides across the dataset and gets the training start and stop indices, plus the index for the associated target. 

# Function to obtain training indices
def get_cutoff_indices(
    data: pd.DataFrame,
    n_features: int, 
    step_size: int) -> list:
    
    stop_position = len(data) - 1
    
    # start first slice at index 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []
    
    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))
        
        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size
        
    return indices
    

# Test function
n_features = 24
step_size = 1

indices = get_cutoff_indices(
    data=ts_data_one_location,
    n_features=n_features,
    step_size=step_size
)

indices[:5]

# generate training data
import numpy as np

n_examples = len(indices)

x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape=n_examples, dtype=np.float32)
pickup_hours = []

for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]]['rides']
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

# Check output is what we want. 
print(f'{x.shape=}')
print(f'{x=}')
print(f'{pickup_hours[:5]=}')

# Now convert to a dataframe
features_one_location = pd.DataFrame(
    x,
    columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
)

features_one_location.head()

# Convert target to dataframe
targets_one_location = pd.DataFrame(
    y,
    columns=['target_rides_next_hour']
)

targets_one_location.head()

# Function to do this for all locations
from tqdm import tqdm

def transform_ts_into_features_and_targets(
    ts_data: pd.DataFrame,
    n_features: int,
    step_size: int  
) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format to (features, target) for training
    """
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}
    
    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        
        # keep only data for this location_id
        ts_data_one_location = ts_data.loc[
            ts_data['pickup_location_id'] == location_id,
            ['pickup_hour', 'rides']
        ]
        
        # pre-compute cutoff indices to split dataframe rows
        indices = get_cutoff_indices(
            ts_data_one_location,
            n_features,
            step_size
        )
        
        # Slice and transpose data into np arrays
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
        y = np.ndarray(shape=n_examples, dtype=np.float32)
        pickup_hours = []

        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1]]['rides']
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])
            
        # features np -> pd
        features_one_location = pd.DataFrame(
            x,
            columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
            )
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id
        
        # target np -> pd
        targets_one_location = pd.DataFrame(
            y,
            columns=['target_rides_next_hour']
            )
        
        # concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])
        
    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True) 
    
    return features, targets['target_rides_next_hour']        

features, targets = transform_ts_into_features_and_targets(
    ts_data,
    n_features=24*7*1, # One week of history
    step_size=24 # to speed things up
)

print(f"{features.shape=}")
print(f"{targets.shape=}")
