==> notebooks/06_baseline_model.ipynb <==
%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df.head()

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df=df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour',
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

# Create first baseline model
class BaselineModelPreviousHour():
    """
    Prediction for target hour = demand from previous hour.
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        return X_test['rides_previous_1_hour']

model = BaselineModelPreviousHour()

predictions = model.predict(X_test)
predictions

# Check predictions against real values for baseline number 1
from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f"{test_mae=:.4f}")

# Create baseline model 2

class BaselineModelPreviousWeek:
    """
    Prediction for target hour t = demand observed at t - 7 days
    e.g. if we want to predict Friday at 6pm, we look at Friday at 6pm the previous week.
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        return X_test[f'rides_previous_{7*24}_hour']

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)

test_mae_baseline_2 = mean_absolute_error(y_test, predictions)
print(f"{test_mae_baseline_2=:.4f}")

# Create baseline model 3
class BaselineModelLast4Weeks:
    """
    Prediction for target hour t = average of demand observed at t - 7 days, t - 14 days, t - 21 days and t - 28 days
    e.g. if we want to predict Friday at 6pm, we look at Friday at 6pm for the previous four weeks and take the average.
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        return 0.25 * (
            X_test[f'rides_previous_{7*24*1}_hour'] + \
            X_test[f'rides_previous_{7*24*2}_hour'] + \
            X_test[f'rides_previous_{7*24*3}_hour'] + \
            X_test[f'rides_previous_{7*24*4}_hour']
        )

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)

test_mae_baseline_3 = mean_absolute_error(y_test, predictions)
print(f"{test_mae_baseline_3=:.4f}")
